requirements.txt
flask==2.3.3
easyocr==1.7.0
numpy==1.24.3
websockets==11.0.3
torch==2.0.1
torchvision==0.15.2
pillow==9.5.0
opencv-python-headless==4.7.0.72
scipy==1.10.1
pyclipper==1.3.0.post4
shapely==2.0.1 